+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Reference creation in case the line has Unicode characters does the replacing correctly by depending on the character count instead of byte count.
+- Backlinks will properly resolve header.
 
 ## [v3.14.4](https://github.com/obsidian-nvim/obsidian.nvim/releases/tag/v3.14.4) - 2025-11-06